requirements.txt
Flask==3.0.2
line-bot-sdk==3.9.0
python-dotenv==1.0.1